[REFACTOR] Finished wrapping component for protected routes
