+++ src/App.js
@@ -1,113 +1,89 @@
-import { useContext } from 'react';
+// import { useContext } from 'react';
 import { Switch, Route, Redirect } from 'react-router-dom';
 import Layout from './components/Layout/Layout';
-import UserProfile from './components/Profile/UserProfile';
 import AuthPage from './pages/AuthForm/AuthPage';
 // import HomePage from './pages/HomePage';
 import UserProfile from './pages/UserProfile/UserProfile';
-import AuthPage from './pages/AuthPage';
-import HomePage from './pages/HomePage';
-import AuthContext from './store/auth-context';
-import Loggedin from './pages/Loggedin';
-import AddProduct from './components/Profile/AddProduct';
-import Inventory from './components/Profile/Inventory';
+// import HomePage from './pages/HomePage';
+// import AuthContext from './store/auth-context';
+import Loggedin from './pages/Loggedin/Userloggedin';
 import EmployeeDisplay from './pages/EmployeeDetails/EmployeeDetails';
-import RevenueReport from './components/Profile/RevenueReport';
-import JobCard from './components/Profile/JobCard';
-import EstimationPage from './pages/Estimation/EstimationPage';
-import Invoice from './components/Profile/invoice';
+import EstimationPage from './pages/Estimation/EstimationDetails';
 import AddProduct from './pages/AddProduct/AddProduct';
 import Inventory from './pages/ProductDetails/Inventory';
-import EmployeeDisplay from './pages/EmployeeDisplay';
 import RevenueReport from './pages/RevenueReport/RevenueReport';
 import JobCard from './pages/JobCard/JobCard';
-import EstimationPage from './pages/EstimationPage';
 import Invoice from './pages/BillGeneration/invoice';
-import SheetDisplay from './components/EmployeeDetails/EmployeeDisplay/SheetDisplay';
+import SheetDisplay from './pages/EmployeeDisplay/SheetDisplay';
+import StartingPageContent from './pages/StartingPage/StartingPageContent';
+import Wrapper from './pages/Wrapper/Wrapper';
+
 
 
 function App() {
-  const authCtx = useContext(AuthContext);
+  // const authCtx = useContext(AuthContext);
 
   return (
     <Layout>
-      <Switch>
-
-      {authCtx.isLoggedIn && ( 
-      <Route path="/employee/managerslist" component={SheetDisplay}/>
-      )}
-
-        {/* {!authCtx.isLoggedIn && (
-          <Route path='/' exact>
-            <HomePage />
+      <Route path='/' exact>
+            <StartingPageContent />
           </Route>
-        )} */}
+      <Wrapper>
+      <Switch>
+      
 
-        {authCtx.isLoggedIn && (
+      <Route path="/employee/managerslist" component={SheetDisplay}/>
+       
+          
 
           <Route path='/' exact>
             <Loggedin />
           </Route>
-        )}
-
-        {!authCtx.isLoggedIn && (
+        
+        
           <Route path='/auth'>
             <AuthPage />
           </Route>
-        )}
+        
         <Route path='/profile'>
-          {authCtx.isLoggedIn &&
-            <UserProfile />}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <UserProfile />
         </Route>
 
         <Route path='/addProduct'>
-          {authCtx.isLoggedIn &&
-            <AddProduct />}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <AddProduct />
         </Route>
-        <Route path='/inventory'>
-          {authCtx.isLoggedIn &&
-            <Inventory />}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+
+        <Route path='/inventory'> 
+            <Inventory />
           </Route>
 
-
-        
         <Route path='/employeeDetails'>
-          {authCtx.isLoggedIn &&
-            <EmployeeDisplay />}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <EmployeeDisplay />
         </Route>
 
         <Route path='/revenuereport'>
-          {authCtx.isLoggedIn &&
-            <RevenueReport />}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <RevenueReport />
         </Route>
 
         <Route path='/jobcard'>
-          {authCtx.isLoggedIn &&
-            <JobCard/>}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <JobCard/>
         </Route>
 
         <Route path='/estimationDetails'>
-          {authCtx.isLoggedIn &&
-            <EstimationPage/>}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <EstimationPage/>
         </Route>
 
         <Route path='/invoice'>
-          {authCtx.isLoggedIn &&
-            <Invoice/>}
-          {!authCtx.isLoggedIn && <Redirect to='/auth' />}
+            <Invoice/>
         </Route>
 
         <Route path='*'>
           <Redirect to='/' />
         </Route>
+       
       </Switch>
+      
+      </Wrapper>
     </Layout>
   );
 }
