+++ src/pages/AuthForm/AuthPage.js
@@ -1,4 +1,5 @@
-import React, { useState,useRef,useContext,AuthContext } from 'react';
+import React, { useState,useRef,useContext, } from 'react';
+import AuthContext from '../../store/auth-context';
 import classes from './AuthPage.module.css';
 
 const AuthPage = () => {
