--- src/pages/Loggedin.js
@@ -1,7 +1,0 @@
-import Userloggedin from "../components/Loggedin/Userloggedin";
-
-const Loggedin = () => {
-  return <Userloggedin />;
-};
-
-export default Loggedin;
